Moved WorkingDictionary to its own file
