+++ Source/Frontend/UI/UI_Extensions.cs
@@ -1,16 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Drawing;
-using System.Globalization;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Windows.Forms;
-using static RTCV.UI.UI_Extensions;
-
 #pragma warning disable RCS1138 // Add summary element to documentation comment.
 #pragma warning disable SA1615
 #pragma warning disable SA1514
@@ -19,6 +6,18 @@ using static RTCV.UI.UI_Extensions;
 
 namespace RTCV.UI
 {
+    using System;
+    using System.Collections.Generic;
+    using System.ComponentModel;
+    using System.Diagnostics;
+    using System.Drawing;
+    using System.Globalization;
+    using System.Linq;
+    using System.Runtime.InteropServices;
+    using System.Text;
+    using System.Windows.Forms;
+    using static RTCV.UI.UI_Extensions;
+
     public static class UI_Extensions
     {
         private static NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
@@ -2065,34 +2064,4 @@ namespace RTCV.UI
             }
         }
     }
-}
-
-//From bizhawk
-/// <summary>
-/// A dictionary that creates new values on the fly as necessary so that any key you need will be defined.
-/// </summary>
-/// <typeparam name="TKey">dictionary keys</typeparam>
-/// <typeparam name="TValue">dictionary values</typeparam>
-[Serializable]
-public class WorkingDictionary<TKey, TValue> : Dictionary<TKey, TValue>
-    where TValue : new()
-{
-    public new TValue this[TKey key]
-    {
-        get
-        {
-            if (!TryGetValue(key, out TValue temp))
-            {
-                temp = this[key] = new TValue();
-            }
-
-            return temp;
-        }
-
-        set => base[key] = value;
-    }
-
-    public WorkingDictionary() { }
-
-    protected WorkingDictionary(SerializationInfo info, StreamingContext context) : base(info, context) { }
 }
